Rename notifications folder
