+++ src/app/LoggedInUseCaseFactory.java
@@ -24,9 +24,9 @@ import use_case.location_lookup.LocationLookupOutputBoundary;
 import use_case.login_signup_switch.LoginSignupSwitchInputBoundary;
 import use_case.login_signup_switch.LoginSignupSwitchInteractor;
 import use_case.login_signup_switch.LoginSignupSwitchOutputBoundary;
-import use_case.notifacation.NotificationInputBoundary;
-import use_case.notifacation.NotificationInteractor;
-import use_case.notifacation.NotificationOutputBoundary;
+import use_case.notification.NotificationInputBoundary;
+import use_case.notification.NotificationInteractor;
+import use_case.notification.NotificationOutputBoundary;
 import use_case.settings_pressed.SettingsPressedInputBoundary;
 import use_case.settings_pressed.SettingsPressedInteractor;
 import use_case.settings_pressed.SettingsPressedOutputBoundary;
