+++ src/use_case/notification/NotificationInteractor.java
@@ -1,4 +1,4 @@
-package use_case.notifacation;
+package use_case.notification;
 
 public class NotificationInteractor implements NotificationInputBoundary {
     final NotificationOutputBoundary presenter;
