+++ src/use_case/notification/NotificationOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.notifacation;
+package use_case.notification;
 
 public interface NotificationOutputBoundary {
     void prepareSuccessView(NotificationOutputData notification);
