+++ src/use_case/notification/NotificationUserDataAccessInterface.java
@@ -1,4 +1,4 @@
-package use_case.notifacation;
+package use_case.notification;
 
 public interface NotificationUserDataAccessInterface {
